[tracker] add ellipse and track id

+++ src/util/bbox_utils.py
@@ -1,6 +1,6 @@
 def get_center_bbox(bbox):
     x1, y1, x2, y2 = bbox
-    return int((x1 + x2) / 2, (y1 + y2) / 2)
+    return int((x1 + x2) / 2), int((y1 + y2) / 2)
 
 def get_bbox_width(bbox):
     x1, y1, x2, y2 = bbox
